refactor(seo): extract costOfLiving test exports to sibling file (file-size guideline)
